offset,limitのクエリパラメータのバリデーションを共通化
