+++ app/controllers/api/v1/tweets_controller.rb
@@ -3,9 +3,7 @@
 module Api
   module V1
     class TweetsController < ApplicationController
-      DEFAULT_LIMIT = 20
-      DEFAULT_OFFSET = 0
-      MAX_LIMIT = 20
+      include PaginationParamsValidator
 
       before_action :authenticate_api_v1_user!, except: %i[index show]
       before_action :validate_query_params, only: %i[index]
@@ -39,26 +37,6 @@ module Api
 
       def tweet_params
         params.require(:tweet).permit(:body)
-      end
-
-      def validate_query_params
-        limit = params[:limit] || DEFAULT_LIMIT.to_s
-        offset = params[:offset] || DEFAULT_OFFSET.to_s
-
-        render_invalid_params and return unless valid_integer_params?(limit, offset)
-
-        @limit = limit.to_i
-        @offset = offset.to_i
-
-        render_invalid_params unless @limit.between?(0, MAX_LIMIT)
-      end
-
-      def valid_integer_params?(limit, offset)
-        limit =~ /\A\d+\z/ && offset =~ /\A\d+\z/ # 正の整数
-      end
-
-      def render_invalid_params
-        render json: { errors: [I18n.t('query_parameters.tweets.invalid_query_parameters')] }, status: :bad_request
       end
     end
   end
